Refactor Vms
